+++ orury-client/src/main/java/org/orury/client/auth/application/oauth/OAuthService.java
@@ -1,0 +1,7 @@
+package org.orury.client.auth.application.oauth;
+
+public interface OAuthService {
+    String getEmailFromOAuthCode(String oAuthCode);
+
+    int getSignUpType();
+}
