--- orury-client/src/main/java/org/orury/client/auth/strategy/LoginStrategy.java
@@ -1,10 +1,0 @@
-package org.orury.client.auth.strategy;
-
-import org.orury.client.auth.converter.request.LoginRequest;
-import org.orury.domain.auth.dto.LoginDto;
-
-public interface LoginStrategy {
-    LoginDto login(LoginRequest request);
-
-    int getSignUpType();
-}
